third_party/WebKit/LayoutTests/fast/canvas/canvas-fillText-zeroSizeGradient.html

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<body>
<script>
test(function(t) {
    var ctx = document.createElement('canvas').getContext('2d');
    
    ctx.fillStyle = '#0f0';
    ctx.fillRect(0, 0, 1, 1);
    
    var g = ctx.createLinearGradient(0, 0, 0, 0); // zero-length line (undefined direction);
    g.addColorStop(0, '#f00');
    g.addColorStop(1, '#f00');
    ctx.fillStyle = g;
    ctx.font = '1px sans-serif';
    ctx.fillText("AA", 0, 1);
    
    var imageData = ctx.getImageData(0, 0, 1, 1);
    var imgdata = imageData.data;
    assert_equals(imgdata[0], 0);
    assert_equals(imgdata[1], 255);
    assert_equals(imgdata[2], 0);
    
}, 'Series of tests to ensure that fillText() paints nothing on canvas when the fillStyle is set to a zero-size gradient.');
</script>
</body>
